Fix all Rust compilation errors and clippy warnings
- Fix exif::Tag GPS_GPS* -> GPS* (remove GPS_ prefix) in import.rs, map.rs
- Fix infer mime_type() returns &str, not Option (and_then -> map)
- Fix exif Value::Ascii bytes type Vec<Vec<u8>> (use first()) in import.rs, map.rs
- Fix compression.rs Option<&str> vs Option<String> mismatch in or_else
- Fix sync.rs u64 -> u32 for total_files field
- Fix pqc.rs: context() -> map_err on hkdf::InvalidLength
- Fix pqc.rs: SigningKey/VerifyingKey to_bytes -> inherent to_seed/encode
- Fix pqc.rs: SigningKey from_bytes -> KeyInit::new_from_slice, add KeyInit import
- Fix pqc.rs: KEM from_bytes() returns Result, add map_err unwraps
- Fix pqc.rs: HmacSha512::new_from_slice ambiguity with fully qualified syntax
- Fix pqc.rs: ArrayString<64> vs String comparison with .as_str()
- Fix encryption.rs: actual_algorithm_used double move by extracting variable
- Fix faces.rs: E0507 cannot move out of Vec index
- Fix faces.rs: E0502 ft_table borrow conflicts (clone data before insert)
- Fix web_dashboard/mod.rs: E0597 table lifetime in delete_by_id
- Fix tantivy_index.rs: replace terms_for_field with per-segment term iteration
- Fix Sidebar.vue orphaned CSS properties causing WASM build failure
- Remove unused imports (redb::ReadableTable) and prefix unused variables

+++ src-tauri/src/commands/sync.rs
@@ -140,7 +140,7 @@ pub fn start_sync(
     {
         let mut progress = sync_state.progress.lock().map_err(|e| e.to_string())?;
         *progress = SyncProgress {
-            total_files: file_ids.len() as u64,
+            total_files: file_ids.len() as u32,
             processed_files: 0,
             current_file: Some("Starting sync...".to_string()),
             status: SyncStatus::Syncing,
